move components from types to ./src
